zetcd: increment zxid on path validation failures
This commit implements the PerfectZXidMode functionality for failed path
validation, causing the ZXid of the server to be incremented on path validation
failures.

This brings the server in-line with upstream Zookeeper, which increments the
ZXid *before* potentially failing the request due to a bad path.

Building with PerfectZXidMode set to false turns the behavior into a noop.