Use last Value for get instead of third Value
Sometimes (such as when doing a get on /), the server's response won't
have any Kvs in the second response. This causes zetcd to crash when a
client does a get on /. I'm not certain this change is valid, but it
does prevent the crash. It also gives some rather arbitrary looking data
for the node's value, probably because it's using an arbitrary response
instead of the hard-coded second response. Maybe we need to always use
Responses[2], but if it has no Kvs, use an empty byte array instead of
crashing?