testing: repair docker cross check tests
The xchk with docker integration tests are broken.  The docker files were
attempting to download old versions of Kafka and drill that no longer
existed on the servers.  In addition the RUOK test was causing all tests
to hang in the docker mode by causing a NPE Zookeeper and never
completely starting up.